requirements.txt
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
filetype==1.0.5
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
har2case==0.3.1
httprunner==2.2.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
locustio==0.11.0
MarkupSafe==1.1.1
msgpack==0.6.1
pycparser==2.19
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
requests-toolbelt==0.8.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.6
